Addressed review comments: updated the custom section in unt test itself
